record.html with watchPosition and setinterval
i replaced the getCurrentPosition in startrecord function to
watchPosition. seem like it can get the more accurate location
And i tried to using loop for keep to running the function, but it was
delayed when it getting the GPS location
I will keep trying my best. check it out if anything have to change